+++ src/picounits/configuration/management.py
@@ -15,6 +15,7 @@ from typing import Dict, Any
 from picounits.configuration.picounits import DEFAULT_ORDER, DEFAULT_SYMBOLS
 from picounits.configuration.picounits import DEFAULT_SIGNIFICANT_FIGURES
 
+# pylint: disable=line-too-long
 
 # Effective preferences after first load
 _effective_symbols: Dict[str, str] | None = None
@@ -27,21 +28,21 @@ _effective_derived: Dict[str, Any] = {}
 
 def get_base_symbols() -> Dict[str, str]:
     """ Gets the base symbol from config """
-    if _effective_symbols is None: _load_config()
+    if _effective_symbols is None: _load_config(None)
 
     return _effective_symbols
 
 
 def get_base_order() -> Dict[str, int]:
     """ Gets the base order from config """
-    if _effective_order is None: _load_config()
+    if _effective_order is None: _load_config(None)
 
     return _effective_order
 
 
 def get_significant_figures() -> int:
     """ Gets the significant figures """
-    if _effective_figures is None: _load_config()
+    if _effective_figures is None: _load_config(None)
 
     return _effective_figures
 
@@ -51,31 +52,37 @@ def get_derived_units() -> Dict[str, Any]:
     return _effective_derived
 
 
-def reload_config() -> None:
-    """ reloads configuration """
-    global _effective_symbols, _effective_order, _effective_figures
-    _effective_symbols, _effective_order, _effective_figures = None, None, None
+def inject_unit_frame(filepath: Path | str) -> None:
+    """ Injects a unit frame for applications from path """
 
-    _load_config()
+    # Converts filepath to Path and Checks file type
+    path = Path(filepath)
+    if path.suffix.lower() != '.picounits':
+        msg = f"Expected .picounits file, got {path.suffix}"
+        raise ImportError(msg) from None
+
+    _load_config(filepath)
 
 
-def _load_config() -> None:
+def _load_config(filepath: Path | None = None) -> None:
     """ Loads the configuration """
     global _effective_symbols, _effective_order, _effective_figures
-    local_file = _find_picounits_file()
 
-    if local_file:
+    if filepath is None:
+        # If no filepath, searches the local dictionary
+        filepath = _find_picounits_file()
+
+    if filepath:
         try:
-            symbols, order, figures = _load_from_file(local_file)
+            symbols, order, figures = _load_from_file(filepath)
             _effective_symbols = {**DEFAULT_SYMBOLS, **symbols}
             _effective_order = order
             _effective_figures = figures
             return
 
         except Exception as e:
-            raise RuntimeError(
-                f"picounits: Failed to parse {local_file}, using defaults: {e}"
-            ) from e
+            msg = f"picounits: Failed to parse {filepath}, using defaults: {e}"
+            raise RuntimeError(msg) from e
 
     # No file or failed use defaults
     _effective_symbols = DEFAULT_SYMBOLS.copy()
@@ -165,10 +172,7 @@ def add_derived_units(registry: Dict[str, Any]) -> None:
         return
 
     if _effective_derived:
-        msg = (
-            "Only one .ut file can be imported at once. "
-            f"Already contains {len(_effective_derived)} units."
-        )
+        msg = f"Only one .ut file can be imported at once. Already contains {len(_effective_derived)} units."
         raise RuntimeError(msg)
 
     if not registry: